Fix condition in cluster state checks
Conditions in Ansible "when" clauses are combined with "and". But the
cluster check should fail if only one of these conditions is true.

Additionally the second condition on the node status check was inversed.
It checked for a healthy node.

Change-Id: Ib2cd632f437c4771452c22f77bedd3012f37b86b
(cherry picked from commit da6abc6d0af2d8f082652d4040894f2ef8958250)